backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
pymongo==4.7.3
python-dotenv==1.0.1
requests==2.32.3
gunicorn==22.0.0
